Every function that do generic type reflexivity is renamed "generic*" and has an "erased*" counterpart.
All functions are now inside the kodein-core module. The kodein module defines extension functions that alias to the  "generic*" functions by default.
The kodein-erased module is the same, but with functions that alias to the "erased*" functions.

Renamed typeToken into genericToken.

In Android, deprecated the scope functions as they have to make a choice (they choose generic*).